requirements.txt
Django==2.2.5
django-bootstrap-modal-forms==1.4.4
django-crispy-forms==1.7.2
django-widget-tweaks==1.4.5
Pillow==6.1.0
pytz==2019.2
sqlparse==0.3.0
